feat: Handle lack of IPv4 address gracefully
If IPv4 address is not found on a node, the application panics:

```
goroutine 1 [running]:
github.com/gathertown/casper-3/pkg/kubernetes.(*Cluster).Nodes(0xc0003f3d20, 0x0, 0x0, 0xc00021bed8, 0xa, 0xa)
	/app/pkg/kubernetes/nodes.go:30 +0x2ae
main.main()
	/app/cmd/casper-3/main.go:53 +0x4d5

```

This PR handles this situation gracefully. If the IPv4 address is not
found, we're not adding the node the list of existing nods in the
kubernetes cluster.