feat: Add prometheus metrics
Adding the prometheus exporter to the app. Introducing two custom metrics:

- Errors encountered
- Total amount of records on each zone and provider

The "errors" encountered assumes that execution keeps running,
otherwise the counter will be reset to zero before prometheus can scrape
it. Maybe this is not the case for _every_ error.

We are interested mainly in errors encountered in the
provider pkgs.