feat: support AWS node names
The AWS node name is `ip-1-2-3-4.ec.internal` which breaks convention.
Add an extra check to look for a dot in the hostname. If a dot is found,
then split and fetch only the first part.

Additionally the `ExternalIP` is found in position 3 of the `Addresses`
array in DigitalOcean while it's in position 2 on AWS. Using a smarter
way to find out `ExternalIP`.