feat: Offload label filtering on CloudFlare
Currently, in order to find which DNS Records exist per cluster, we
are fetching all TXT records from CloudFlare, and then we perform the
label filtering on our side. This leads to quite expensive requests,
since we don't really need to fetch all TXT records of our zone. We
can optimise our queries to CloudFlare and reduce our code complexity
by offloading the label filtering on the Cloudflare[1].
This commit enables this functionality.

1. https://github.com/cloudflare/cloudflare-go/issues/613#issuecomment-813747784