feat:Compare record & node prefixes before deletion
This commit introduces a common function `RecordPrefixMatchesNodePrefixes`
that aims to check if there is a match between record and node
prefixes. Casper-3 should be allowed to delete only records that
are prefixed with the labeled k8s node prefixes too. Let's add this
check as a safeguard here, in order to avoid the deletion of unwilling
records.