feat: Validate record name & type before deletion
Records should be validated before the deletion. Casper-3 is allowed
to delete only records that are type `A` or `TXT` and their name is
the same with the fqdn input of the deleteRecord() function. In case
that the deleteRecord function processes records that don't fill these
requirements, the deletion should not happen