[circleci] Switch to doctl orb
The doks orb has the tendency to install too many unecessary tools like
kubectl and kops. We do not really need that as we installing kubectl
through the official orb and we also do not need kops. As such lets
switch to doctl orb instead which only installs the doctl tool.