feat: Adjust casper-3 limits
Adjust casper-3 limits and requests to make space for other pods, in
this case falco-IDS. Although we do not define `requests`, kubernetes
will use the `limits` to set `requests` as well.

Casper-3 uses very low vCPU (max is 75m according to Prometheus in
sgp-1b) and memory rarely goes above 32MB.

Use these as requests and keep the limits close as we don't see spikes
in a 14d span.