fix: empty CLOUDFLARE_PROXY_POOLS broken parsing
Split function in the `strings` package, `func Split(s, sep string) []string` has a behaviour where
if s does not contain sep and sep is not empty, Split returns a slice of length 1 whose only element is s.
This breaks our expectation of having a empty node pool list in case, the env variable is not set.

Handling this case explicitly fixes the issue.